refactor: int 범위 넘어가지 않을 domain PK 자료형 int로 수정

+++ src/main/java/com/newbie/ajou/domain/college/CollegeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CollegeRepository extends JpaRepository<College, Long> {
+public interface CollegeRepository extends JpaRepository<College, Integer> {
 	Optional<College> findByName(String name);
 }
